auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "A service to provide authentication and authorization API"
authors = ["rifatrakib <Rakib.1508@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.101.1"}
python-decouple = "^3.8"
pydantic-settings = "^2.0.3"
pydash = "^7.0.6"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.20"}
asyncpg = "^0.28.0"
alembic = "^1.11.3"
aioredis = "^2.0.1"
fastapi-mail = "^1.4.1"
passlib = {extras = ["bcrypt", "argon2"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
